feat: update files usinstypescript
