+++ nodemon.json
@@ -1,0 +1,5 @@
+{
+  "watch": ["**"],
+  "ext": "ts",
+  "exec": "ts-node src/index.ts"
+}
