+++ tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "compilerOptions": {
+    "module": "es2022",
+    "esModuleInterop": true,
+    "moduleResolution": "Node",
+    "target": "es2017",
+    "outDir": "dist",
+    "baseUrl": "./src",
+  },
+  "include": [
+    "./src/**/*"
+  ],
+}
